+++ Mage/src/main/java/mage/abilities/ActivatedAbilityImpl.java
@@ -29,6 +29,7 @@ package mage.abilities;
 
 import java.util.UUID;
 import mage.MageObject;
+import mage.abilities.condition.Condition;
 import mage.abilities.costs.Cost;
 import mage.abilities.costs.Costs;
 import mage.abilities.costs.mana.ManaCosts;
@@ -45,6 +46,7 @@ import mage.constants.Zone;
 import mage.game.Game;
 import mage.game.command.Emblem;
 import mage.game.permanent.Permanent;
+import mage.util.CardUtil;
 
 /**
  *
@@ -52,6 +54,19 @@ import mage.game.permanent.Permanent;
  */
 public abstract class ActivatedAbilityImpl extends AbilityImpl implements ActivatedAbility {
 
+    static class ActivationInfo {
+
+        public int turnNum;
+        public int activationCounter;
+
+        public ActivationInfo(int turnNum, int activationCounter) {
+            this.turnNum = turnNum;
+            this.activationCounter = activationCounter;
+        }
+    }
+
+    protected int maxActivationsPerTurn = Integer.MAX_VALUE;
+    protected Condition condition;
     protected TimingRule timing = TimingRule.INSTANT;
     protected TargetController mayActivate = TargetController.YOU;
     protected UUID activatorId;
@@ -68,6 +83,8 @@ public abstract class ActivatedAbilityImpl extends AbilityImpl implements Activa
         mayActivate = ability.mayActivate;
         activatorId = ability.activatorId;
         checkPlayableMode = ability.checkPlayableMode;
+        maxActivationsPerTurn = ability.maxActivationsPerTurn;
+        condition = ability.condition;
     }
 
     public ActivatedAbilityImpl(Zone zone) {
@@ -161,6 +178,9 @@ public abstract class ActivatedAbilityImpl extends AbilityImpl implements Activa
     @Override
     public boolean canActivate(UUID playerId, Game game) {
         //20091005 - 602.2
+        if (!(hasMoreActivationsThisTurn(game) && (condition == null || condition.apply(game, this)))) {
+            return false;
+        }
         switch (mayActivate) {
             case ANY:
                 break;
@@ -255,4 +275,46 @@ public abstract class ActivatedAbilityImpl extends AbilityImpl implements Activa
         return checkPlayableMode;
     }
 
+    private boolean hasMoreActivationsThisTurn(Game game) {
+        ActivationInfo activationInfo = getActivationInfo(game);
+        return activationInfo == null || activationInfo.turnNum != game.getTurnNum() || activationInfo.activationCounter < maxActivationsPerTurn;
+    }
+
+    @Override
+    public boolean activate(Game game, boolean noMana) {
+        if (hasMoreActivationsThisTurn(game)) {
+            if (super.activate(game, noMana)) {
+                ActivationInfo activationInfo = getActivationInfo(game);
+                if (activationInfo == null) {
+                    activationInfo = new ActivationInfo(game.getTurnNum(), 1);
+                } else if (activationInfo.turnNum != game.getTurnNum()) {
+                    activationInfo.turnNum = game.getTurnNum();
+                    activationInfo.activationCounter = 1;
+                } else {
+                    activationInfo.activationCounter++;
+                }
+                setActivationInfo(activationInfo, game);
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public void setMaxActivationsPerTurn(int maxActivationsPerTurn) {
+        this.maxActivationsPerTurn = maxActivationsPerTurn;
+    }
+
+    private ActivationInfo getActivationInfo(Game game) {
+        Integer turnNum = (Integer) game.getState().getValue(CardUtil.getCardZoneString("activationsTurn", sourceId, game));
+        Integer activationCount = (Integer) game.getState().getValue(CardUtil.getCardZoneString("activationsCount", sourceId, game));
+        if (turnNum == null || activationCount == null) {
+            return null;
+        }
+        return new ActivationInfo(turnNum, activationCount);
+    }
+
+    private void setActivationInfo(ActivationInfo activationInfo, Game game) {
+        game.getState().setValue(CardUtil.getCardZoneString("activationsTurn", sourceId, game), activationInfo.turnNum);
+        game.getState().setValue(CardUtil.getCardZoneString("activationsCount", sourceId, game), activationInfo.activationCounter);
+    }
 }
