+++ Mage/src/main/java/mage/abilities/common/LimitedTimesPerTurnActivatedAbility.java
@@ -36,27 +36,11 @@ import mage.constants.Zone;
 import mage.game.Game;
 import mage.util.CardUtil;
 
-import java.util.UUID;
-
 /**
  *
  * @author BetaSteward_at_googlemail.com
  */
 public class LimitedTimesPerTurnActivatedAbility extends ActivatedAbilityImpl {
-
-    static class ActivationInfo {
-
-        public int turnNum;
-        public int activationCounter;
-
-        public ActivationInfo(int turnNum, int activationCounter) {
-            this.turnNum = turnNum;
-            this.activationCounter = activationCounter;
-        }
-    }
-
-    private int maxActivationsPerTurn;
-    private Condition condition;
 
     public LimitedTimesPerTurnActivatedAbility(Zone zone, Effect effect, Cost cost) {
         this(zone, effect, cost, 1);
@@ -76,38 +60,6 @@ public class LimitedTimesPerTurnActivatedAbility extends ActivatedAbilityImpl {
         super(ability);
         this.maxActivationsPerTurn = ability.maxActivationsPerTurn;
         this.condition = ability.condition;
-    }
-
-    @Override
-    public boolean canActivate(UUID playerId, Game game) {
-        return super.canActivate(playerId, game)
-                && hasMoreActivationsThisTurn(game)
-                && (condition == null || condition.apply(game, this));
-    }
-
-    private boolean hasMoreActivationsThisTurn(Game game) {
-        ActivationInfo activationInfo = getActivationInfo(game);
-        return activationInfo == null || activationInfo.turnNum != game.getTurnNum() || activationInfo.activationCounter < maxActivationsPerTurn;
-    }
-
-    @Override
-    public boolean activate(Game game, boolean noMana) {
-        if (hasMoreActivationsThisTurn(game)) {
-            if (super.activate(game, noMana)) {
-                ActivationInfo activationInfo = getActivationInfo(game);
-                if (activationInfo == null) {
-                    activationInfo = new ActivationInfo(game.getTurnNum(), 1);
-                } else if (activationInfo.turnNum != game.getTurnNum()) {
-                    activationInfo.turnNum = game.getTurnNum();
-                    activationInfo.activationCounter = 1;
-                } else {
-                    activationInfo.activationCounter++;
-                }
-                setActivationInfo(activationInfo, game);
-                return true;
-            }
-        }
-        return false;
     }
 
     @Override
@@ -141,19 +93,5 @@ public class LimitedTimesPerTurnActivatedAbility extends ActivatedAbilityImpl {
     @Override
     public LimitedTimesPerTurnActivatedAbility copy() {
         return new LimitedTimesPerTurnActivatedAbility(this);
-    }
-
-    private ActivationInfo getActivationInfo(Game game) {
-        Integer turnNum = (Integer) game.getState().getValue(CardUtil.getCardZoneString("activationsTurn", sourceId, game));
-        Integer activationCount = (Integer) game.getState().getValue(CardUtil.getCardZoneString("activationsCount", sourceId, game));
-        if (turnNum == null || activationCount == null) {
-            return null;
-        }
-        return new ActivationInfo(turnNum, activationCount);
-    }
-
-    private void setActivationInfo(ActivationInfo activationInfo, Game game) {
-        game.getState().setValue(CardUtil.getCardZoneString("activationsTurn", sourceId, game), activationInfo.turnNum);
-        game.getState().setValue(CardUtil.getCardZoneString("activationsCount", sourceId, game), activationInfo.activationCounter);
     }
 }
